+++ src/App/user_profile.resolver.ts
@@ -15,14 +15,14 @@ class UserProfileResolver {
 
   @Query(() => UserProfile)
   async getProfile(
-    @Args('id', { type: () => String, nullable: true, defaultValue: '' })
+    @Args('id', { type: () => String, nullable: true})
     id: string,
-    @Args('username', { type: () => String, nullable: true, defaultValue: '' })
+    @Args('username', { type: () => String, nullable: true })
     username: string,
   ) {
     const repository = this.connection.getRepository(UserProfile)
     return repository.findOneOrFail({
-      where: `LOWER(id) = '${id.toLowerCase()}' OR LOWER(username) = '${username.toLowerCase()}'`,
+      where: `LOWER(id) = '${id?.toLowerCase()}' OR LOWER(username) = '${username?.toLowerCase()}'`,
     })
   }
